# Optimize Variadic Elementwise Ops